# perf(core): improve hash table lookup performance for small string keys such as column names and symbols