--- a/core/src/main/java/io/questdb/cairo/map/Unordered2Map.java
+++ b/core/src/main/java/io/questdb/cairo/map/Unordered2Map.java
@@ -140,7 +140,7 @@ public class Unordered2Map implements Map, Reopenable {
     }
 
     @Override
-    public final void close() {
+    public void close() {
         if (memStart != 0) {
             memStart = memLimit = Unsafe.free(memStart, entrySize * TABLE_SIZE, memoryTag);
             keyMemStart = Unsafe.free(keyMemStart, KEY_SIZE, memoryTag);