--- a/core/src/main/java/io/questdb/cairo/map/Unordered8MapRecord.java
+++ b/core/src/main/java/io/questdb/cairo/map/Unordered8MapRecord.java
@@ -290,8 +290,8 @@ final class Unordered8MapRecord implements MapRecord {
     }
 
     @Override
-    public int keyHashCode() {
-        return Hash.hashLong(Unsafe.getUnsafe().getLong(startAddress));
+    public long keyHashCode() {
+        return Hash.hashLong64(Unsafe.getUnsafe().getLong(startAddress));
     }
 
     public void of(long address) {