Reduces the number of hash code collisions in hash tables used in table/column name caches, especially for strings with differences in a few last characters.

Includes the following changes:
* Decreases load factors in hash tables used for table/column name caches.
* Adds `Hash#spread()` finalizer to `Utf8StringIntHashMap` and `Utf8StringObjHashMap`. Perf improvement shown below comes mainly from this change.
* Adds MurmurHash finalizer to hash functions from `Hash`.
* Converts hash functions used in `Map` implementations to 64-bit. This helps with less collisions in sharded map scenarios, i.e. when there are many groups aggregated by a GROUP BY query.

### Microbenchmarks

```
Benchmark                                                 (len)  Mode  Cnt    Score    Error  Units
StringHashFunctionBenchmark.testHashUtf8                      7  avgt    3    4.914 ±  1.876  ns/op
StringHashFunctionBenchmark.testHashUtf8                     15  avgt    3    6.151 ±  2.879  ns/op
StringHashFunctionBenchmark.testHashUtf8                     31  avgt    3    8.328 ±  0.529  ns/op
StringHashFunctionBenchmark.testHashUtf8                     63  avgt    3   12.032 ±  2.327  ns/op
StringHashFunctionBenchmark.testHashUtf8                   1024  avgt    3  138.408 ±  0.742  ns/op
StringHashFunctionBenchmark.testStandardHashCharSequence      7  avgt    3    6.219 ±  1.614  ns/op
StringHashFunctionBenchmark.testStandardHashCharSequence     15  avgt    3   11.333 ±  0.684  ns/op
StringHashFunctionBenchmark.testStandardHashCharSequence     31  avgt    3   24.067 ±  0.899  ns/op
StringHashFunctionBenchmark.testStandardHashCharSequence     63  avgt    3   53.197 ±  0.736  ns/op
StringHashFunctionBenchmark.testStandardHashCharSequence   1024  avgt    3  907.643 ± 14.018  ns/op
```

`Utf8StringIntHashMapBenchmark` uses string keys with differences in a few last characters. The `prefixLength` parameter specifies the length of the identical part of the strings.
```
Benchmark                                    (n)  (prefixLength)  Mode  Cnt      Score      Error  Units
Utf8StringIntHashMapBenchmark.testUtf16Map    16               5  avgt    3    118.189 ±    6.163  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map    16               7  avgt    3    144.288 ±    6.233  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map    16              31  avgt    3    438.458 ±    9.096  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map    16              63  avgt    3    915.738 ±   27.589  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map   256               5  avgt    3   2001.784 ±   39.868  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map   256               7  avgt    3   2255.600 ±  914.338  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map   256              31  avgt    3   6850.364 ±  108.525  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map   256              63  avgt    3  14577.537 ±  135.006  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map  1024               5  avgt    3   9047.163 ±  961.982  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map  1024               7  avgt    3   9742.841 ± 5326.609  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map  1024              31  avgt    3  28918.299 ± 8820.928  ns/op
Utf8StringIntHashMapBenchmark.testUtf16Map  1024              63  avgt    3  58900.686 ± 2143.095  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map     16               5  avgt    3    104.839 ±   31.743  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map     16               7  avgt    3     79.055 ±   62.847  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map     16              31  avgt    3    136.725 ±   21.795  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map     16              63  avgt    3    198.512 ±  100.215  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map    256               5  avgt    3   1303.660 ±  432.664  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map    256               7  avgt    3   1493.835 ±  397.566  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map    256              31  avgt    3   2258.106 ±  283.167  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map    256              63  avgt    3   3465.830 ± 1492.208  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map   1024               5  avgt    3   5207.690 ± 3090.980  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map   1024               7  avgt    3   8100.634 ± 6388.336  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map   1024              31  avgt    3  10954.549 ±  149.008  ns/op
Utf8StringIntHashMapBenchmark.testUtf8Map   1024              63  avgt    3  15563.850 ± 1083.949  ns/op
```

### Benchmarks

Modified TSBS: https://github.com/questdb/tsbs/tree/1k_column_table_experiment (table with 1k columns)

Before:
```bash
$ cat /tmp/qdb-data-1k.gz | gunzip | ./tsbs_load_questdb --workers=2
time,per. metric/s,metric total,overall metric/s,per. row/s,row total,overall row/s
1711643508,5939411.24,5.940000E+07,5939411.24,5999.41,6.000000E+04,5999.41
1711643518,7920592.07,1.386000E+08,6929915.53,8000.60,1.400000E+05,6999.91
1711643528,6930020.65,2.079000E+08,6929950.57,7000.02,2.100000E+05,6999.95
1711643538,6930002.83,2.772000E+08,6929963.64,7000.00,2.800000E+05,6999.96
1711643548,5940010.18,3.366000E+08,6731974.05,6000.01,3.400000E+05,6799.97

Summary:
loaded 356400000 metrics in 51.342sec with 2 workers (mean rate 6941630.71 metrics/sec)
loaded 360000 rows in 51.342sec with 2 workers (mean rate 7011.75 rows/sec)
```

After:
```bash
$ cat /tmp/qdb-data-1k.gz | gunzip | ./tsbs_load_questdb --workers=2
time,per. metric/s,metric total,overall metric/s,per. row/s,row total,overall row/s
1711643394,13857949.49,1.386000E+08,13857949.49,13997.93,1.400000E+05,13997.93
1711643404,14850021.75,2.871000E+08,14353948.56,15000.02,2.900000E+05,14498.94

Summary:
loaded 356400000 metrics in 24.073sec with 2 workers (mean rate 14804847.93 metrics/sec)
loaded 360000 rows in 24.073sec with 2 workers (mean rate 14954.39 rows/sec)
```