@ideoma 2024-03-28
#### [PR Coverage check]

:heart_eyes: **pass** : 132 / 135 (97.78%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/str/DirectUtf8Sink.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/cairo/map/ShardedMapCursor.java|1|2|50.00%|
|:large_blue_circle:|io/questdb/std/Utf8StringObjHashMap.java|6|7|85.71%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctUuidGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered4Map.java|7|7|100.00%|
|:large_blue_circle:|io/questdb/std/Utf8StringIntHashMap.java|9|9|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered8Map.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/OrderedMapFixedSizeRecord.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByLongHashSet.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByLong128HashSet.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/std/Hash.java|22|22|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered8MapRecord.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered4MapRecord.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByAtom.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/std/AbstractLongHashSet.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered16Map.java|7|7|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByIntHashSet.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/std/Uuid.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByRecordCursorFactory.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctIPv4GroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctLongGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/OrderedMap.java|29|29|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered16MapRecord.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/std/LongLongHashSet.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/OrderedMapVarSizeRecord.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByLong256HashSet.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctLong256GroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/std/ConcurrentLongHashMap.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/std/CharSequenceIntHashMap.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctIntGroupByFunction.java|1|1|100.00%|